Initial import of idGrab
This is a still heavily in-progress tool to generate headers for
id-Engine games (e.g., Keen 4–6) from a script file. The script file is
not in IGRAB format, and the tool does not grab actual graphics data
(though that could be added later if I felt like it), it's just a list
of chunks and properties.

At the moment, idGrab can generate:
- Omnispeak-compatible GFXCHUNK.* variables.
- TED5-compatible GFXINFOE.* binary headers.
- modId / uGrab compatible .def files.
- IGRAB-style C/ASM headers {GRAPH*,GFXE_*}.{H,EQU}

It's written in horribly non-idiomatic Rust, and only depends on the
standard library (core, std), and (optionally) libc for an IGRAB-style
timestamp in the header.

See some sample scripts for Keen: Galaxy in the scripts/ directory.

Example use:
- igrab scripts/keen4.idgrab --igrab-header gfxe_ck4.h --igrab-asm
  gfxe_ck4.equ
- igrab scripts/keen5.igrab --gfxinfo GFXINFOE.CK5
- cargo run scripts/keen6.igrab --omnispeak GFXCHUNKS.CK6

Enjoy!
— David