Add --igrab-version option, fix DEMO lines in ASM headers
You can now set the IGRAB version emulated by the --igrab-header and
--igrab-asm options, to use 0.24 or 0.4 (default). The --igrab-version
option must be passed in _before_ the corresponding --igrab-{header,asm}
option (this allows you to make headers for both versions).

The differences are:
- 0.24 uses #defines, 0.4 uses an enum
- 0.4 includes MISC chunks.
- Note that IGRAB uses GFXE_[EXT].{H,EQU} as the filename in 0.4, and
  GRAPH[EXT].{H,EQU} for 0.24.
- Versions 0.4 is available, with source, as part of the TED5
  distribution. Version 0.24 was used to generate Keen Dreams' data.
  (The source was never released. The binary has been preserved, but is
  not publicly available).

We also fix an erroneous comma in the ASM output after DEMO lines. There
are still a few minor formatting issues, but it otherwise works okay
here.