Fix timestamps feature on architectures/ABIs with signed char
Rust is a bit confused about when to use std::ffi::c_char and when to
use u8 for C strings, meaning that the libc crate's ctime_r() function
wants a C-style 'char' array (which on one of my machines is signed),
but std::ffi::CStr::from_bytes_until_nul() wants a u8, which is always
unsigned.

Fortunately, we can work around this with a cast, but I think that
std::ffi::CStr should support c_char slices — working with C strings is,
after all, what it's for.