igrab: Add --igrab-underscore-separator option
This inserts an '_' between a chunk name and its suffix. Keep in mind
that this is experimental: I may remove it, make it only change sprites,
rename it, etc.

The purpose here is to allow idGrab to generate headers compatible with
K1n9_Duk3's decompilation and lethal-guitar's Bio Menace decompilation,
which use an '_SPR' separator. Of course, this doesn't work yet, as they
don't use a separator for the PIC and PICM suffixes. But the refactoring
of IGrabOptions will be useful regardless.