Support games missing some header chunks
Most igrab games have at least some bitmaps, masked bitmaps, and sprites, the
details of which are stored in the first three "header chunks". However, if
a game doesn't, then there should be fewer than three chunks reserved as
headers.

This is needed, for example, by Wolfenstein 3D and friends, which do not have
any masked bitmaps or sprites, so should only have one header chunk (for
unmasked bitmaps), and have the fonts start at chunk number 1.

There's probably some further improvement to the code here we could do to
avoid hardcoding the number of headers, and instead calculate it when required.
(Or at least verify it matches).